+++ R/choices.R
@@ -2,6 +2,7 @@
 #'
 #' Simulate choices for a survey design, either randomly or according to a
 #' utility model defined by user-provided prior parameters.
+#' @keywords logitr mnl mxl mixed logit simulation
 #'
 #' @param design A data frame of a survey design.
 #' @param obsID The name of the column in `design` that identifies each choice
@@ -140,6 +141,8 @@ def_model_prior <- function(design, priors, n_draws) {
     parSetup     = parSetup,
     parIDs       = parIDs,
     standardDraws = getStandardDraws(parIDs, n_draws),
+    # Create data object
+    data = list(factorLevels = codedData$factorLevels),
     # Create n object, which stores counts of various variables
     n = list(
       vars       = length(parSetup),