+++ NEWS.md
@@ -1,3 +1,15 @@
+# cbcTools (development version)
+
+
+# cbcTools 0.2.0
+
+- Added a startup message when the package is loaded
+- Updates for compatibility with logitr version 1.0.1.
+
+# cbcTools 0.1.0
+
+- Added initial integration with {idefix} packages for Bayesian D-efficient designs
+
 # cbcTools 0.0.5
 
 - Updates for compatibility with logitr version 0.8.0.